+++ src/OpenTelemetry.Exporter.Stackdriver/Implementation/ActivityExtensions.cs
@@ -14,8 +14,6 @@
 // limitations under the License.
 // </copyright>
 
-#nullable enable
-
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
@@ -136,25 +134,25 @@ internal static class ActivityExtensions
         switch (av)
         {
             case string s:
-                return new AttributeValue()
+                return new AttributeValue
                 {
-                    StringValue = new TruncatableString() { Value = s },
+                    StringValue = new TruncatableString { Value = s },
                 };
             case bool b:
-                return new AttributeValue() { BoolValue = b };
+                return new AttributeValue { BoolValue = b };
             case long l:
-                return new AttributeValue() { IntValue = l };
+                return new AttributeValue { IntValue = l };
             case double d:
-                return new AttributeValue()
+                return new AttributeValue
                 {
-                    StringValue = new TruncatableString() { Value = d.ToString(CultureInfo.InvariantCulture) },
+                    StringValue = new TruncatableString { Value = d.ToString(CultureInfo.InvariantCulture) },
                 };
             case null:
                 return new AttributeValue();
             default:
-                return new AttributeValue()
+                return new AttributeValue
                 {
-                    StringValue = new TruncatableString() { Value = av.ToString() },
+                    StringValue = new TruncatableString { Value = av.ToString() },
                 };
         }
     }