+++ test/OpenTelemetry.Sampler.AWS.Tests/TestAWSXRaySamplerClient.cs
@@ -17,7 +17,6 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
-using OpenTelemetry.Sampler.AWS;
 using WireMock.RequestBuilders;
 using WireMock.ResponseBuilders;
 using WireMock.Server;